TidalScale/performance-meter/requirements.txt
confluent_kafka==1.9.0
kafka_python==2.0.2
numpy==1.22.4
psycopg2==2.9.5
psycopg2_binary==2.9.3
PyYAML==6.0
